[skip ci] Drop Mentions of EasyBot.py in Description
The project has not been updated in months now, and I have refactored a lot of the code to work better. Thus I'm dropping the part it says "EasyBot.py plugin integration" due to Rin now being more focused on the cogs that I made, and not as heavily influenced as EasyBot.py anymore.