Reuse AIOHTTP + GQL client sessions throughout the lifecycle of Kumiko
When reused, the client sessions become much more efficient. It's the same idea behind the Redis connection pooling system, and this speeds up the commands that pull from APIs by a ton. It's also generally good practice, and the client sessions will automatically close when the bot is being shut down