Fix prefix module + create lookup tables
The lookup tables have indexes on them to help with performance. This idea was ported from Akari, which in turn was borrowed from RoboDanny. In this case, we have indexed the names of the items in the economy. There is no need for a marketplace table, so we can remove that later