Remove unban and only accept one user
The rationale made is that once a user is banned, that user should be banned for good. There exists no purpose of creating an unban command when you quite literally just gave the keys to the troll again.

As a result, it is best to keep that person banned for good. If server admins need to unban a user for some reason, then they should do so manually